Replace the `class ArgumentsLexerBuilder` and `class ArgumentsLexer` interface with one function
Bug: b/389804379
Test: /usr/bin/bazel run :cvd -- --instance_name=1 status
